BUG: Images, fossil links, and user mentions fixed
WordPress' hook system can be tough where there are a ton of hooks and no
common place to see which hooks are enabled or in what order they will be
executed...

Anyway, so we had some of our hooks out-of-order and made the assumption that
the URL expansion to a link hook was disabled, turns out it wasn't--so we were
double parsing a lot of links.

This also broke the behavior for how embedded images were handled, which is
also fixed by this commit.

Closes #180